backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai==1.12.0
hypothesis==6.98.0
dj-database-url==2.1.0
